# perf(loader): improve language lookup speed