### Problem

While running a `release-dev` build to test changes at a faster rate, I noticed that the `parse` subcommand felt like it was hanging for a little bit. Turns out, it was around ~100ms after checking with `hyperfine`:
![image](https://github.com/user-attachments/assets/99250dfa-904c-46d2-a09b-fb81ca83ee6c)


 I decided to poke around and see what was causing the slowdown with a flamegraph, and got something that looked like this:
 
![image](https://github.com/user-attachments/assets/11eb801a-e8ee-4184-a936-3cbb448a5cd8)

It becomes pretty evident that the majority of the time spent here is in serde's `deserialize`, which is called in `find_language_configurations_at_path` in the loader crate. If we take a look, we see there are three spots we use serde's deserialize: [here](https://github.com/tree-sitter/tree-sitter/blob/fe92e978f922b7c8caf8d43604d39f419f963749/cli/loader/src/lib.rs#L1136), [here](https://github.com/tree-sitter/tree-sitter/blob/fe92e978f922b7c8caf8d43604d39f419f963749/cli/loader/src/lib.rs#L1148), and [here](https://github.com/tree-sitter/tree-sitter/blob/fe92e978f922b7c8caf8d43604d39f419f963749/cli/loader/src/lib.rs#L1243).

The first case is not the problem, since we are loading a relatively small file, but the other two are loading a grammar.json file, and in fact, it's the same file being deserialized *twice*. It becomes pretty clear that the first one isn't needed, we deserialize `grammar.json` just to read the name field when we've already successfully parser a config file - we should use the name in the config file instead.

However, this does not solve the ~100ms avg parse subcommand time, because, as one would expect, release-like profiles likely (don't quote me) optimize the two deserialize calls into one as an optimization, since it's the same path we're reading from. So, the real problem is that we're reading potentially thousands of lines long grammar files *just* to extract the field name, and nothing else.

### Solution

Instead of reading & deserializing thousands of lines into memory to extract one field, we can rely on the invariant that if a grammar.json was generated with tree-sitter cli (why wouldn't it be? ;) ), the name field is *always* within the first three lines. We can just read three lines and extract the name field with a simple regex. The results are great: with `release-dev`, the `parse` subcommand goes down from 100ms to 2.3ms, and is just under 2ms with the `release` profile on a Ryzen 9 9950x:
![image](https://github.com/user-attachments/assets/3d5995aa-76bd-4f83-80da-202bfba49d96)
![image](https://github.com/user-attachments/assets/01e88182-3dd2-4494-a7c6-433015684b88)

Another small finding that was observed when playing around with serde's deserialization is that `serde_json::from_reader` with a `BufReader` is a bit slower than `serde_json::from_str` with `fs::read_to_string`. In my testing, for `release-dev` it was ~3.5ms, compared to 2.3ms without it, so every usage of `from_reader` + `BufReader` was replaced with `from_str` + `fs::read_to_string`.