@clason 2024-10-12
This broke `tree-sitter build` for a bunch of grammars:
```
[nvim-treesitter/install/glimmer_typescript] error: Error during "tree-sitter build": thread 'main' panicked at cli/src/main.rs:776:18:
called `Result::unwrap()` on an `Err` value: Failed to parse the language name from grammar.json at /Users/clason/.cache/nvim/tree-sitter-glimmer_typescript/src/grammar.json
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

[nvim-treesitter/install/glimmer_javascript] error: Error during "tree-sitter build": thread 'main' panicked at cli/src/main.rs:776:18:
called `Result::unwrap()` on an `Err` value: Failed to parse the language name from grammar.json at /Users/clason/.cache/nvim/tree-sitter-glimmer_javascript/src/grammar.json
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
...
[nvim-treesitter/install/slang] error: Error during "tree-sitter build": thread 'main' panicked at cli/src/main.rs:776:18:
called `Result::unwrap()` on an `Err` value: Failed to parse the language name from grammar.json at /Users/clason/.cache/nvim/tree-sitter-slang/src/grammar.json
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
...
[nvim-treesitter/install/hlsl] error: Error during "tree-sitter build": thread 'main' panicked at cli/src/main.rs:776:18:
called `Result::unwrap()` on an `Err` value: Failed to parse the language name from grammar.json at /Users/clason/.cache/nvim/tree-sitter-hlsl/src/grammar.json
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
...
[nvim-treesitter/install/qmljs] error: Error during "tree-sitter build": thread 'main' panicked at cli/src/main.rs:776:18:
called `Result::unwrap()` on an `Err` value: Failed to parse the language name from grammar.json at /Users/clason/.cache/nvim/tree-sitter-qmljs/src/grammar.json
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```
(Due to missing "inherits" field, from a quick sample.)